Moved files to their corresponding folder, changed css styles to maintain consistency
