+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './logo.svg';
 import { Counter } from './features/counter/Counter';
 import './App.css';
-import Registration from './UI/UIRegistration';
+import Registration from './UI/registration/UIRegistration';
 
 function App() {
   return (
