+++ src/Components/formik/Input.js
@@ -3,11 +3,10 @@ import { Field, ErrorMessage } from 'formik'
 import TextError from './TextError'
 
 function Input (props) {
-  const { label, name, ...rest } = props
+  const { place_holder,label, name, ...rest } = props
   return (
     <div className='form-control'>
-      <label htmlFor={name}>{label}</label>
-      <Field id={name} name={name} {...rest} />
+      <Field id={name} name={name} {...rest} placeholder={place_holder} />
       <ErrorMessage component={TextError} name={name} />
     </div>
   )
