+++ src/UI/registration/CSSRegistration.css
@@ -39,7 +39,7 @@ label {
   .submit_btn{
     width: 100%;
     height: 3em;
-    background-color: red;
+    background-color: #DB5752;
     color: white;
     font-weight: bold;
   }
